refactor: improve error handling and exception management

+++ src/main/java/com/tomachocolate/api/service/ParticipantService.java
@@ -1,5 +1,6 @@
 package com.tomachocolate.api.service;
 
+import com.tomachocolate.api.exception.BadRequestException;
 import com.tomachocolate.api.exception.ResourceNotFoundException;
 import com.tomachocolate.api.model.Meeting;
 import com.tomachocolate.api.model.Participant;
@@ -24,17 +25,24 @@ public class ParticipantService {
         Meeting meeting = meetingRepository.findById(meetingId)
                 .orElseThrow(() -> new ResourceNotFoundException("Juntada no encontrada"));
 
-        Participant p = new Participant();
-        p.setName(name);
-        meeting.addParticipant(p);
-        meeting.setParticipantCount(meeting.getParticipantCount() +1);
-        return participantRepository.save(p);
+        if(meeting.getParticipantCount() == 30){
+            throw new BadRequestException("Máximo de 30 participantes");
         }
+
+        if(name == null || name.isBlank()){
+            throw new BadRequestException("Ingrese un nombre válido");
+        } else {
+            Participant p = new Participant();
+            p.setName(name);
+            meeting.addParticipant(p);
+            meeting.setParticipantCount(meeting.getParticipantCount() +1);
+            return participantRepository.save(p);
+        }}
 
     @Transactional
     public void updateName(Long id, String newName) {
         Participant participant = participantRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("No pudimos encontrar al participante con ID: " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("El participante no existe"));
 
         participant.setName(newName);
         participantRepository.save(participant);
@@ -43,13 +51,17 @@ public class ParticipantService {
     @Transactional
     public void deleteParticipant(Long id) {
         Participant participant = participantRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("No pudimos encontrar al participante con ID: " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("El participante no existe"));
 
         Meeting meeting = participant.getMeeting();
 
-        expenseRepository.deleteByPayer(participant);
-        participantRepository.delete(participant);
-        meeting.setParticipantCount(meeting.getParticipantCount() - 1);
+        if (meeting.getParticipantCount() <= 2) {
+            throw new BadRequestException("Mínimo 2 participantes");
+        } else {
+            expenseRepository.deleteByPayer(participant);
+            participantRepository.delete(participant);
+            meeting.setParticipantCount(meeting.getParticipantCount() - 1);
+        }
     }
 
 }
